Fix always true condition

+++ src/server/game/Movement/Spline/MoveSpline.cpp
@@ -236,10 +236,10 @@ MoveSpline::UpdateResult MoveSpline::_updateState(uint32& ms_time_diff)
 
     UpdateResult result = Result_None;
 
-    uint32 minimal_diff = std::min(ms_time_diff, segment_time_elapsed());
+    int32 minimal_diff = std::min(ms_time_diff, segment_time_elapsed());
     ASSERT(minimal_diff >= 0);
-    time_passed += minimal_diff;
-    ms_time_diff -= minimal_diff;
+    time_passed += uint32(minimal_diff);
+    ms_time_diff -= uint32(minimal_diff);
 
     if (time_passed >= next_timestamp())
     {
